+++ ChangeLog.rst
@@ -1,3 +1,9 @@
+Unreleased Changes
+------------------
+
+* Truly disable the writeback cache, instead of just adjusting the
+  default and printing a warning when explicitly enabled.
+
 Release 3.1.0 (2017-08-04)
 --------------------------